Extracted plugin list in own view and made tabs into sidebar
