+++ futuremod/src/view/dashboard/components.rs
@@ -1,34 +1,32 @@
-use std::collections::HashMap;
-
-use futuremod_data::plugin::{Plugin, PluginDependency, PluginState};
-use iced::{alignment::{Horizontal, Vertical}, widget::{center, column, container, mouse_area, opaque, row, rule, scrollable, text, Space, Stack, Toggler}, Alignment, Length, Padding};
-use iced_aw::Badge;
+use futuremod_data::plugin::PluginDependency;
+use iced::{alignment::{Horizontal, Vertical}, widget::{center, column, container, mouse_area, opaque, row, rule, scrollable, text, Space, Stack}, Alignment, Length, Padding};
 use iced_fonts::Bootstrap;
-use log::warn;
 
-use crate::{theme::{self, Container, Theme}, widget::{bold, button, icon_button, icon_text_button, icon_text_button_advanced, Column, Element, IconTextButtonOptions, Row}};
+use crate::{theme::{self, Container, Theme}, widget::{button, icon_button, icon_text_button, Column, Element}};
 
 use super::{view::{Dialog, InstallConfirmationPrompt, View}, Dashboard, Message};
 
 pub fn dashboard<'a>(state: &'a Dashboard) -> Element<'a, Message> {
   let content = match &state.view {
-    None => plugin_overview(&state.plugins, state.is_developer),
-    Some(view) => match view {
-      View::Logs(logs_view) => logs_view.view(&state.logs).map(Message::Logs),
-      View::Plugin(plugin_view) => {
-        match state.plugins.get(&plugin_view.name) {
-          Some(plugin) => plugin_view.view(&plugin).map(Message::Plugin),
-          None => error_box("Plugin doesn't exist".to_string()),
-        }
+    View::Logs(logs_view) => logs_view.view(&state.logs).map(Message::Logs),
+    View::Plugin(plugin_view) => {
+      match state.plugins.get(&plugin_view.name) {
+        Some(plugin) => plugin_view.view(&plugin).map(Message::Plugin),
+        None => error_box("Plugin doesn't exist".to_string()),
       }
     },
+    View::PluginList(plugin_list_view) => plugin_list_view.view(&state.plugins, state.is_developer)
+      .map(Message::PluginList),
   };
 
   let underlay: Element<Message> = column![
-    heading(state.is_developer, &state.view),
+    heading(state.is_developer),
     rule::Rule::horizontal(1.0),
-    Space::new(Length::Fill, 8.0),
-    content
+    row![
+      sidebar(&state.view),
+      rule::Rule::vertical(1.0),
+      content,
+    ]
   ].into();
 
   let mut overlay: Option<Element<Message>> = None;
@@ -52,6 +50,14 @@ pub fn dashboard<'a>(state: &'a Dashboard) -> Element<'a, Message> {
     .push_maybe(overlay)
     .width(Length::Fill)
     .height(Length::Fill)
+    .into()
+}
+
+fn sidebar(active_view: &View) -> Element<'_, Message> {
+  container(
+    tabs(active_view)
+  )
+    .padding(16)
     .into()
 }
 
@@ -107,14 +113,11 @@ fn dialog_header<'a>(title: String) -> Element<'a, Message> {
     .into()
 }
 
-fn heading<'a>(is_developer: bool, active_view: &Option<View>) -> Element<'a, Message> {
+fn heading<'a>(is_developer: bool) -> Element<'a, Message> {
   row![
     container(title(is_developer)).width(Length::Fill).padding(Padding{top: 0.0, right: 0.0, bottom: 0.0, left: 8.0}),
-    container(
-      tabs(active_view)
-    )
-      .padding(8.0),
   ]
+    .padding(8)
     .align_y(Alignment::Center)
     .into()
 }
@@ -128,184 +131,30 @@ fn title<'a>(is_developer: bool) -> Element<'a, Message> {
   text(content).size(32.0).into()
 }
 
-fn tabs<'a>(active_view: &Option<View>) -> Element<'a, Message> {
-  let is_plugin_tab = |active_view: &Option<View>| {
+fn tabs<'a>(active_view: &View) -> Element<'a, Message> {
+  let is_plugin_tab = |active_view: &View| {
     match active_view {
-      None | Some(View::Plugin(_)) => true,
+      View::PluginList(_) | View::Plugin(_) => true,
       _ => false,
     }
   };
 
-  row![
-    icon_text_button(Bootstrap::Box, "Plugins")
-      .on_press(Message::ToPlugins)
-      .class(if is_plugin_tab(&active_view) {theme::Button::Primary} else {theme::Button::HoverHighlight}),
-    icon_text_button(Bootstrap::CardText, "Logs")
-      .on_press(Message::ToLogs)
-      .class(if let Some(View::Logs(_)) = active_view {theme::Button::Primary} else {theme::Button::HoverHighlight}),
-    icon_text_button(Bootstrap::Gear, "Settings")
-      .on_press(Message::ToSettings),
-  ]
-    .spacing(8.0)
-    .into()
-}
+  let tab_button = |icon: Bootstrap, label: &'static str, on_press: Option<Message>, active: bool| -> Element<'_, Message> {
+    icon_text_button(icon, label)
+      .on_press_maybe(on_press)
+      .class(if active {theme::Button::Primary} else {theme::Button::HoverHighlight})
+      .width(200)
+      .into()
+  };
 
-fn plugin_overview<'a>(plugins: &'a HashMap<String, Plugin>, is_developer: bool) -> Element<'a, Message> {
   column![
-    plugin_overview_heading(is_developer),
-    plugin_list(plugins),
+    tab_button(Bootstrap::Box, "Plugins", Some(Message::ToPluginList), is_plugin_tab(&active_view)),
+    tab_button(Bootstrap::CardText, "Logs", Some(Message::ToLogs), matches!(active_view, View::Logs(_))),
+    Space::with_height(Length::Fill),
+    tab_button(Bootstrap::Gear, "Settings", None, false),
   ]
-    .spacing(16)
-    .padding(16)
-    .into()
-}
-
-fn plugin_overview_heading<'a>(is_developer: bool) -> Element<'a, Message> {
-  row![
-    container(text("Plugins").size(24)).width(Length::Fill),
-    plugin_overview_actions(is_developer),
-  ]
-    .width(Length::Fill)
-    .into()
-}
-
-fn plugin_overview_actions<'a>(is_developer: bool) -> Element<'a, Message> {
-  let mut actions = Row::new();
-
-  if is_developer {
-    actions = actions.push(
-      icon_text_button(Bootstrap::Bug, "Install as Developer")
-        .class(theme::Button::Default)
-        .on_press(Message::StartDevelopmentInstallation)
-    );
-  }
-
-  actions = actions.push(
-    icon_text_button_advanced(Bootstrap::Plus, "Install", IconTextButtonOptions::new().with_icon_size(24))
-      .class(theme::Button::Primary)
-      .padding(Padding{top: 3.0, right: 16.0, bottom: 3.0, left: 12.0})
-      .on_press(Message::StartInstallation)
-  );
-
-  actions
     .spacing(8.0)
     .into()
-}
-
-fn plugin_list<'a>(plugins: &'a HashMap<String, Plugin>) -> Element<'a, Message> {
-  let mut list = Column::new();
-
-  let mut keys: Vec<&String> = plugins.keys().collect();
-  keys.sort();
-
-  for name in keys {
-    match plugins.get(name) {
-      Some(plugin) => {
-        list = list.push(plugin_card(name, plugin));
-      },
-      None => {
-        warn!("Missing plugin while generating plugin list");
-      }
-    }
-  }
-
-  list
-    .spacing(16.0)
-    .into()
-}
-
-fn plugin_card<'a>(name: &'a String, plugin: &Plugin) -> Element<'a, Message> {
-  container(
-    row![
-      Column::new()
-        .push(text(name).size(20))
-        .push(plugin_card_description(&plugin))
-        .width(Length::Fill)
-        .spacing(8.0),
-      plugin_card_actions(&plugin),
-    ]
-    .align_y(Alignment::Center)
-  )
-  .class(Container::Box)
-  .padding(16)
-  .into()
-}
-
-fn plugin_card_actions<'a>(plugin: &Plugin) -> Element<'a, Message> {
-  Row::new()
-    .push(plugin_go_to_details_button(&plugin))
-    .push_maybe(plugin_toggle_button(&plugin))
-    .spacing(8)
-    .align_y(Alignment::Center)
-    .into()
-}
-
-fn plugin_card_description<'a>(plugin: &Plugin) -> Element<'a, Message> {
-  let dev_mode_badge: Option<Element<'a, Message>> = match &plugin.in_dev_mode {
-    false => None,
-    true => Some(Badge::new(text("Developer Mode").font(bold()).size(12))
-      .style(|theme: &theme::Theme, status| iced_aw::style::badge::warning(&theme.theme, status))
-      .padding(4)
-      .into()),
-  };
-
-  Row::new()
-    .push(plugin_state_component(plugin))
-    .push_maybe(dev_mode_badge)
-    .align_y(Alignment::Center)
-    .spacing(4.0)
-    .into()
-}
-
-fn plugin_state_component<'a>(plugin: &Plugin) -> Element<'a, Message> {
-  let message = match &plugin.state {
-    PluginState::Error(_) => String::from("Error"),
-    PluginState::Unloaded => String::from("Unloaded"),
-    _ => {
-      if plugin.enabled {
-        String::from("Enabled")
-      } else {
-        String::from("Disabled")
-      }
-    }
-  };
-
-  text(message)
-    .size(14)
-    .into()
-}
-
-fn plugin_go_to_details_button<'a>(plugin: &Plugin) -> Element<'a, Message> {
-  button(text("Details"))
-    .on_press(Message::ToPlugin(plugin.info.name.clone()))
-    .class(theme::Button::Default)
-    .into()
-}
-
-fn plugin_toggle_button<'a>(plugin: &Plugin) -> Option<Element<'a, Message>> {
-  if let PluginState::Error(_) = plugin.state {
-    return None;
-  }
-
-  let label = match plugin.enabled {
-    true => "Enabled",
-    false => "Disabled",
-  };
-
-  let plugin_name = plugin.info.name.clone();
-  let enabled = plugin.enabled;
-
-  Some(
-    container(
-      Toggler::new(enabled)
-        .label(label)
-        .on_toggle(move |state| match state {
-          true => Message::Enable(plugin_name.clone()),
-          false => Message::Disable(plugin_name.clone()),
-        })
-        .width(120)
-  ).into()
-  )
 }
 
 fn dependencies_list<'a>(dependencies: &Vec<PluginDependency>) -> Element<'a, Message> {
