+++ futuremod/src/view/mod.rs
@@ -2,3 +2,4 @@ pub mod loading;
 pub mod logs;
 pub mod dashboard;
 pub mod plugin;
+pub mod plugin_list;
