+++ futuremod/src/view/plugin_list/mod.rs
@@ -1,0 +1,5 @@
+pub mod components;
+pub mod state;
+pub mod view;
+
+pub use view::{PluginList, Message};
